feat: rewrite cjsToEsm.cjs
